@@ -3,7 +3,7 @@ description: Calorie Difference indicator for AppleHealth users.
 publish_to: 'none'
 version: 1.5.2
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.8.0 <4.0.0"
 
 dependencies:
   device_check: 1.0.0
@@ -11,36 +11,36 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  google_fonts: 6.2.1
-  health: 10.2.0
-  flutter_hooks: 0.20.5
-  flutter_riverpod: 2.5.1
-  hooks_riverpod: 2.5.1
-  introduction_screen: 3.1.14
-  shared_preferences: 2.2.3
+  google_fonts: ^6.3.0
+  health: 13.1.1
+  flutter_hooks: ^0.21.3
+  flutter_riverpod: ^2.6.1
+  hooks_riverpod: ^2.6.1
+  introduction_screen: ^3.1.17
+  shared_preferences: 2.5.3
   cupertino_icons: 1.0.8
-  intl: 0.19.0
-  localizely_sdk: 2.5.7
-  rate_my_app: 2.2.0
-  freezed: 2.5.3
-  json_serializable: 6.8.0
-  firebase_core: 2.32.0
-  firebase_analytics: 10.10.7
-  firebase_crashlytics: 3.5.7
+  intl: 0.20.2
+  localizely_sdk: 2.7.1
+  rate_my_app: ^2.3.2
+  freezed: 3.2.0
+  json_serializable: 6.10.0
+  firebase_core: 4.0.0
+  firebase_analytics: 12.0.0
+  firebase_crashlytics: 5.0.0
 
 dev_dependencies:
-  build_runner: 2.4.10
-  mockito: 5.4.4
+  build_runner: 2.7.0
+  mockito: 5.5.0
   remove_from_coverage: 2.0.0
   flutter_test:
     sdk: flutter
-  flutter_lints: 4.0.0
+  flutter_lints: 6.0.0
+  frontend_server_client: 4.0.0
 
 dependency_overrides:
-  meta: 1.15.0
-#  uuid: 4.4.0
-#  http: 1.2.1
-
+  build: 3.0.2
+  build_runner_core: 9.3.0
+  win32: 5.5.1
 
 flutter:
   uses-material-design: true
@@ -54,5 +54,4 @@ cider:
 flutter_intl:
   enabled: true
   localizely:
-    ota_enabled: true
     project_id: acdfa863-c3ee-42c6-a9e5-da9c1eee7dae
