@@ -1,11 +1,8 @@
-import 'dart:io';
-
 import 'package:calorie_diff/providers/macros_providers.dart';
 import 'package:calorie_diff/theme/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:localizely_sdk/localizely_sdk.dart';
 
 import 'core/core_providers.dart';
 import 'generated/l10n.dart';
@@ -15,7 +12,7 @@ import 'providers/calories_providers.dart';
 part 'info_popup.dart';
 
 class CalorieDiffApp extends StatefulHookConsumerWidget {
-  const CalorieDiffApp({Key? key}) : super(key: key);
+  const CalorieDiffApp({super.key});
 
   @override
   ConsumerState<CalorieDiffApp> createState() => _CalorieDiffAppState();
@@ -29,9 +26,6 @@ class _CalorieDiffAppState extends ConsumerState<CalorieDiffApp>
   void initState() {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
-    if (!Platform.environment.containsKey('FLUTTER_TEST')) {
-      Localizely.updateTranslations();
-    }
   }
 
   @override
