@@ -3,7 +3,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'health_calories_model.freezed.dart';
 
 @freezed
-class HealthCaloriesModel with _$HealthCaloriesModel {
+abstract class HealthCaloriesModel with _$HealthCaloriesModel {
   const HealthCaloriesModel._();
 
   const factory HealthCaloriesModel({
