@@ -2,11 +2,10 @@ import 'package:calorie_diff/core/extensions.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'health_macros_model.freezed.dart';
-
 part 'health_macros_model.g.dart';
 
 @freezed
-class HealthMacrosModel with _$HealthMacrosModel {
+abstract class HealthMacrosModel with _$HealthMacrosModel {
   const HealthMacrosModel._();
 
   const factory HealthMacrosModel({
@@ -17,11 +16,11 @@ class HealthMacrosModel with _$HealthMacrosModel {
   }) = _HealthMacrosModel;
 
   factory HealthMacrosModel.initial() => HealthMacrosModel(
-        date: ExtendedDateTime.current,
-        carb: 0,
-        fat: 0,
-        protein: 0,
-      );
+    date: ExtendedDateTime.current,
+    carb: 0,
+    fat: 0,
+    protein: 0,
+  );
 
   factory HealthMacrosModel.fromJson(Map<String, dynamic> json) =>
       _$HealthMacrosModelFromJson(json);
