@@ -1,5 +1,5 @@
-// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// coverage:ignore-file
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
@@ -9,177 +9,290 @@ part of 'settings_model.dart';
 // FreezedGenerator
 // **************************************************************************
 
+// dart format off
 T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
-
-SettingsModel _$SettingsModelFromJson(Map<String, dynamic> json) {
-  return _SettingsModel.fromJson(json);
-}
 
 /// @nodoc
 mixin _$SettingsModel {
-  bool get macrosEnabled => throw _privateConstructorUsedError;
-  HealthMacrosModel get macros => throw _privateConstructorUsedError;
 
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
-  $SettingsModelCopyWith<SettingsModel> get copyWith =>
-      throw _privateConstructorUsedError;
+ bool get macrosEnabled; HealthMacrosModel get macros;
+/// Create a copy of SettingsModel
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$SettingsModelCopyWith<SettingsModel> get copyWith => _$SettingsModelCopyWithImpl<SettingsModel>(this as SettingsModel, _$identity);
+
+  /// Serializes this SettingsModel to a JSON map.
+  Map<String, dynamic> toJson();
+
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is SettingsModel&&(identical(other.macrosEnabled, macrosEnabled) || other.macrosEnabled == macrosEnabled)&&(identical(other.macros, macros) || other.macros == macros));
+}
+
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,macrosEnabled,macros);
+
+@override
+String toString() {
+  return 'SettingsModel(macrosEnabled: $macrosEnabled, macros: $macros)';
+}
+
+
 }
 
 /// @nodoc
-abstract class $SettingsModelCopyWith<$Res> {
-  factory $SettingsModelCopyWith(
-          SettingsModel value, $Res Function(SettingsModel) then) =
-      _$SettingsModelCopyWithImpl<$Res, SettingsModel>;
-  @useResult
-  $Res call({bool macrosEnabled, HealthMacrosModel macros});
+abstract mixin class $SettingsModelCopyWith<$Res>  {
+  factory $SettingsModelCopyWith(SettingsModel value, $Res Function(SettingsModel) _then) = _$SettingsModelCopyWithImpl;
+@useResult
+$Res call({
+ bool macrosEnabled, HealthMacrosModel macros
+});
 
-  $HealthMacrosModelCopyWith<$Res> get macros;
+
+$HealthMacrosModelCopyWith<$Res> get macros;
+
 }
-
 /// @nodoc
-class _$SettingsModelCopyWithImpl<$Res, $Val extends SettingsModel>
+class _$SettingsModelCopyWithImpl<$Res>
     implements $SettingsModelCopyWith<$Res> {
-  _$SettingsModelCopyWithImpl(this._value, this._then);
+  _$SettingsModelCopyWithImpl(this._self, this._then);
 
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
+  final SettingsModel _self;
+  final $Res Function(SettingsModel) _then;
 
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? macrosEnabled = null,
-    Object? macros = null,
-  }) {
-    return _then(_value.copyWith(
-      macrosEnabled: null == macrosEnabled
-          ? _value.macrosEnabled
-          : macrosEnabled // ignore: cast_nullable_to_non_nullable
-              as bool,
-      macros: null == macros
-          ? _value.macros
-          : macros // ignore: cast_nullable_to_non_nullable
-              as HealthMacrosModel,
-    ) as $Val);
-  }
-
-  @override
-  @pragma('vm:prefer-inline')
-  $HealthMacrosModelCopyWith<$Res> get macros {
-    return $HealthMacrosModelCopyWith<$Res>(_value.macros, (value) {
-      return _then(_value.copyWith(macros: value) as $Val);
-    });
-  }
+/// Create a copy of SettingsModel
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? macrosEnabled = null,Object? macros = null,}) {
+  return _then(_self.copyWith(
+macrosEnabled: null == macrosEnabled ? _self.macrosEnabled : macrosEnabled // ignore: cast_nullable_to_non_nullable
+as bool,macros: null == macros ? _self.macros : macros // ignore: cast_nullable_to_non_nullable
+as HealthMacrosModel,
+  ));
+}
+/// Create a copy of SettingsModel
+/// with the given fields replaced by the non-null parameter values.
+@override
+@pragma('vm:prefer-inline')
+$HealthMacrosModelCopyWith<$Res> get macros {
+  
+  return $HealthMacrosModelCopyWith<$Res>(_self.macros, (value) {
+    return _then(_self.copyWith(macros: value));
+  });
+}
 }
 
-/// @nodoc
-abstract class _$$SettingsModelImplCopyWith<$Res>
-    implements $SettingsModelCopyWith<$Res> {
-  factory _$$SettingsModelImplCopyWith(
-          _$SettingsModelImpl value, $Res Function(_$SettingsModelImpl) then) =
-      __$$SettingsModelImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call({bool macrosEnabled, HealthMacrosModel macros});
 
-  @override
-  $HealthMacrosModelCopyWith<$Res> get macros;
+/// Adds pattern-matching-related methods to [SettingsModel].
+extension SettingsModelPatterns on SettingsModel {
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _SettingsModel value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _SettingsModel() when $default != null:
+return $default(_that);case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _SettingsModel value)  $default,){
+final _that = this;
+switch (_that) {
+case _SettingsModel():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _SettingsModel value)?  $default,){
+final _that = this;
+switch (_that) {
+case _SettingsModel() when $default != null:
+return $default(_that);case _:
+  return null;
+
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( bool macrosEnabled,  HealthMacrosModel macros)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _SettingsModel() when $default != null:
+return $default(_that.macrosEnabled,_that.macros);case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( bool macrosEnabled,  HealthMacrosModel macros)  $default,) {final _that = this;
+switch (_that) {
+case _SettingsModel():
+return $default(_that.macrosEnabled,_that.macros);case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( bool macrosEnabled,  HealthMacrosModel macros)?  $default,) {final _that = this;
+switch (_that) {
+case _SettingsModel() when $default != null:
+return $default(_that.macrosEnabled,_that.macros);case _:
+  return null;
+
+}
 }
 
-/// @nodoc
-class __$$SettingsModelImplCopyWithImpl<$Res>
-    extends _$SettingsModelCopyWithImpl<$Res, _$SettingsModelImpl>
-    implements _$$SettingsModelImplCopyWith<$Res> {
-  __$$SettingsModelImplCopyWithImpl(
-      _$SettingsModelImpl _value, $Res Function(_$SettingsModelImpl) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? macrosEnabled = null,
-    Object? macros = null,
-  }) {
-    return _then(_$SettingsModelImpl(
-      macrosEnabled: null == macrosEnabled
-          ? _value.macrosEnabled
-          : macrosEnabled // ignore: cast_nullable_to_non_nullable
-              as bool,
-      macros: null == macros
-          ? _value.macros
-          : macros // ignore: cast_nullable_to_non_nullable
-              as HealthMacrosModel,
-    ));
-  }
 }
 
 /// @nodoc
 @JsonSerializable()
-class _$SettingsModelImpl extends _SettingsModel {
-  const _$SettingsModelImpl({required this.macrosEnabled, required this.macros})
-      : super._();
 
-  factory _$SettingsModelImpl.fromJson(Map<String, dynamic> json) =>
-      _$$SettingsModelImplFromJson(json);
+class _SettingsModel extends SettingsModel {
+  const _SettingsModel({required this.macrosEnabled, required this.macros}): super._();
+  factory _SettingsModel.fromJson(Map<String, dynamic> json) => _$SettingsModelFromJson(json);
 
-  @override
-  final bool macrosEnabled;
-  @override
-  final HealthMacrosModel macros;
+@override final  bool macrosEnabled;
+@override final  HealthMacrosModel macros;
 
-  @override
-  String toString() {
-    return 'SettingsModel(macrosEnabled: $macrosEnabled, macros: $macros)';
-  }
+/// Create a copy of SettingsModel
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$SettingsModelCopyWith<_SettingsModel> get copyWith => __$SettingsModelCopyWithImpl<_SettingsModel>(this, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$SettingsModelImpl &&
-            (identical(other.macrosEnabled, macrosEnabled) ||
-                other.macrosEnabled == macrosEnabled) &&
-            (identical(other.macros, macros) || other.macros == macros));
-  }
-
-  @JsonKey(ignore: true)
-  @override
-  int get hashCode => Object.hash(runtimeType, macrosEnabled, macros);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$SettingsModelImplCopyWith<_$SettingsModelImpl> get copyWith =>
-      __$$SettingsModelImplCopyWithImpl<_$SettingsModelImpl>(this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$SettingsModelImplToJson(
-      this,
-    );
-  }
+@override
+Map<String, dynamic> toJson() {
+  return _$SettingsModelToJson(this, );
 }
 
-abstract class _SettingsModel extends SettingsModel {
-  const factory _SettingsModel(
-      {required final bool macrosEnabled,
-      required final HealthMacrosModel macros}) = _$SettingsModelImpl;
-  const _SettingsModel._() : super._();
-
-  factory _SettingsModel.fromJson(Map<String, dynamic> json) =
-      _$SettingsModelImpl.fromJson;
-
-  @override
-  bool get macrosEnabled;
-  @override
-  HealthMacrosModel get macros;
-  @override
-  @JsonKey(ignore: true)
-  _$$SettingsModelImplCopyWith<_$SettingsModelImpl> get copyWith =>
-      throw _privateConstructorUsedError;
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _SettingsModel&&(identical(other.macrosEnabled, macrosEnabled) || other.macrosEnabled == macrosEnabled)&&(identical(other.macros, macros) || other.macros == macros));
 }
+
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,macrosEnabled,macros);
+
+@override
+String toString() {
+  return 'SettingsModel(macrosEnabled: $macrosEnabled, macros: $macros)';
+}
+
+
+}
+
+/// @nodoc
+abstract mixin class _$SettingsModelCopyWith<$Res> implements $SettingsModelCopyWith<$Res> {
+  factory _$SettingsModelCopyWith(_SettingsModel value, $Res Function(_SettingsModel) _then) = __$SettingsModelCopyWithImpl;
+@override @useResult
+$Res call({
+ bool macrosEnabled, HealthMacrosModel macros
+});
+
+
+@override $HealthMacrosModelCopyWith<$Res> get macros;
+
+}
+/// @nodoc
+class __$SettingsModelCopyWithImpl<$Res>
+    implements _$SettingsModelCopyWith<$Res> {
+  __$SettingsModelCopyWithImpl(this._self, this._then);
+
+  final _SettingsModel _self;
+  final $Res Function(_SettingsModel) _then;
+
+/// Create a copy of SettingsModel
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? macrosEnabled = null,Object? macros = null,}) {
+  return _then(_SettingsModel(
+macrosEnabled: null == macrosEnabled ? _self.macrosEnabled : macrosEnabled // ignore: cast_nullable_to_non_nullable
+as bool,macros: null == macros ? _self.macros : macros // ignore: cast_nullable_to_non_nullable
+as HealthMacrosModel,
+  ));
+}
+
+/// Create a copy of SettingsModel
+/// with the given fields replaced by the non-null parameter values.
+@override
+@pragma('vm:prefer-inline')
+$HealthMacrosModelCopyWith<$Res> get macros {
+  
+  return $HealthMacrosModelCopyWith<$Res>(_self.macros, (value) {
+    return _then(_self.copyWith(macros: value));
+  });
+}
+}
+
+// dart format on
