@@ -1,6 +1,7 @@
 import 'package:calorie_diff/core/extensions.dart';
 import 'package:calorie_diff/models/health_calories_model.dart';
 import 'package:calorie_diff/shared/utils/health_utils.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:health/health.dart';
 
@@ -30,6 +31,7 @@ final healthCaloriesProvider = FutureProvider<HealthCaloriesModel>((ref) async {
     clean,
     HealthDataType.DIETARY_ENERGY_CONSUMED,
   );
+  debugPrint('Active: $active, Rest: $rest, Dietary: $dietary');
   final difference = dietary - (active + rest);
   return HealthCaloriesModel(
     date: now,
@@ -41,44 +43,45 @@ final healthCaloriesProvider = FutureProvider<HealthCaloriesModel>((ref) async {
 
 final historicHealthDataProvider =
     FutureProvider.family<List<HealthCaloriesModel>, int>((
-  ref, [
-  int days = 7,
-]) async {
-  final now = ExtendedDateTime.current;
-  final start = now.subtract(Duration(days: days));
-  final end = DateTime(now.year, now.month, now.day);
+      ref, [
+      int days = 7,
+    ]) async {
+      final now = ExtendedDateTime.current;
+      final start = now.subtract(Duration(days: days));
+      final end = DateTime(now.year, now.month, now.day);
 
-  final health = ref.read(healthFactoryProvider);
-  final types = ref.read(healthDataTypesProvider);
+      final health = ref.read(healthFactoryProvider);
+      final types = ref.read(healthDataTypesProvider);
 
-  final data = await health.getHealthDataFromTypes(
-    startTime: DateTime(start.year, start.month, start.day),
-    endTime: end,
-    types: types,
-  );
+      final data = await health.getHealthDataFromTypes(
+        startTime: DateTime(start.year, start.month, start.day),
+        endTime: end,
+        types: types,
+      );
 
-  final clean = Health().removeDuplicates(data);
-  final dates = HealthUtils.getDates(clean);
+      final clean = Health().removeDuplicates(data);
+      final dates = HealthUtils.getDates(clean);
 
-  return dates.map((date) {
-    final active = HealthUtils.prepareDataEntry(
-      HealthUtils.filterByDate(clean, date),
-      HealthDataType.ACTIVE_ENERGY_BURNED,
-    );
-    final rest = HealthUtils.prepareDataEntry(
-      HealthUtils.filterByDate(clean, date),
-      HealthDataType.BASAL_ENERGY_BURNED,
-    );
-    final dietary = HealthUtils.prepareDataEntry(
-      HealthUtils.filterByDate(clean, date),
-      HealthDataType.DIETARY_ENERGY_CONSUMED,
-    );
-    final difference = dietary - (active + rest);
-    return HealthCaloriesModel(
-      date: date,
-      burned: HealthUtils.decimals(active + rest),
-      consumed: HealthUtils.decimals(dietary),
-      difference: HealthUtils.decimals(difference),
-    );
-  }).toList();
-});
+      return dates.map((date) {
+        final active = HealthUtils.prepareDataEntry(
+          HealthUtils.filterByDate(clean, date),
+          HealthDataType.ACTIVE_ENERGY_BURNED,
+        );
+        final rest = HealthUtils.prepareDataEntry(
+          HealthUtils.filterByDate(clean, date),
+          HealthDataType.BASAL_ENERGY_BURNED,
+        );
+        final dietary = HealthUtils.prepareDataEntry(
+          HealthUtils.filterByDate(clean, date),
+          HealthDataType.DIETARY_ENERGY_CONSUMED,
+        );
+        debugPrint('Active: $active, Rest: $rest, Dietary: $dietary');
+        final difference = dietary - (active + rest);
+        return HealthCaloriesModel(
+          date: date,
+          burned: HealthUtils.decimals(active + rest),
+          consumed: HealthUtils.decimals(dietary),
+          difference: HealthUtils.decimals(difference),
+        );
+      }).toList();
+    });
