@@ -13,8 +13,8 @@ class NumericInputField extends HookWidget {
     required this.label,
     required this.callback,
     this.suffixText,
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
