@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.4 from annotations
+// Mocks generated by Mockito 5.4.6 from annotations
 // in calorie_diff/test/mocks/core_mocks.dart.
 // Do not manually edit this file.
 
@@ -12,7 +12,7 @@ import 'package:flutter/src/widgets/scroll_context.dart' as _i8;
 import 'package:flutter/src/widgets/scroll_physics.dart' as _i7;
 import 'package:flutter/src/widgets/scroll_position.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:shared_preferences/shared_preferences.dart' as _i3;
+import 'package:shared_preferences/src/shared_preferences_legacy.dart' as _i3;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -22,6 +22,7 @@ import 'package:shared_preferences/shared_preferences.dart' as _i3;
 // ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: must_be_immutable
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
 // ignore_for_file: camel_case_types
@@ -29,13 +30,8 @@ import 'package:shared_preferences/shared_preferences.dart' as _i3;
 
 class _FakeScrollPosition_0 extends _i1.SmartFake
     implements _i2.ScrollPosition {
-  _FakeScrollPosition_0(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
+  _FakeScrollPosition_0(Object parent, Invocation parentInvocation)
+    : super(parent, parentInvocation);
 }
 
 /// A class which mocks [SharedPreferences].
@@ -43,187 +39,135 @@ class _FakeScrollPosition_0 extends _i1.SmartFake
 /// See the documentation for Mockito's code generation for more information.
 class MockSharedPreferences extends _i1.Mock implements _i3.SharedPreferences {
   @override
-  Set<String> getKeys() => (super.noSuchMethod(
-        Invocation.method(
-          #getKeys,
-          [],
-        ),
-        returnValue: <String>{},
-        returnValueForMissingStub: <String>{},
-      ) as Set<String>);
-  @override
-  Object? get(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #get,
-          [key],
-        ),
-        returnValueForMissingStub: null,
-      ) as Object?);
-  @override
-  bool? getBool(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #getBool,
-          [key],
-        ),
-        returnValueForMissingStub: null,
-      ) as bool?);
-  @override
-  int? getInt(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #getInt,
-          [key],
-        ),
-        returnValueForMissingStub: null,
-      ) as int?);
-  @override
-  double? getDouble(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #getDouble,
-          [key],
-        ),
-        returnValueForMissingStub: null,
-      ) as double?);
-  @override
-  String? getString(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #getString,
-          [key],
-        ),
-        returnValueForMissingStub: null,
-      ) as String?);
-  @override
-  bool containsKey(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #containsKey,
-          [key],
-        ),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
-  @override
-  List<String>? getStringList(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #getStringList,
-          [key],
-        ),
-        returnValueForMissingStub: null,
-      ) as List<String>?);
-  @override
-  _i4.Future<bool> setBool(
-    String? key,
-    bool? value,
-  ) =>
+  Set<String> getKeys() =>
       (super.noSuchMethod(
-        Invocation.method(
-          #setBool,
-          [
-            key,
-            value,
-          ],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
+            Invocation.method(#getKeys, []),
+            returnValue: <String>{},
+            returnValueForMissingStub: <String>{},
+          )
+          as Set<String>);
   @override
-  _i4.Future<bool> setInt(
-    String? key,
-    int? value,
-  ) =>
+  Object? get(String? key) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #setInt,
-          [
-            key,
-            value,
-          ],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
+            Invocation.method(#get, [key]),
+            returnValueForMissingStub: null,
+          )
+          as Object?);
   @override
-  _i4.Future<bool> setDouble(
-    String? key,
-    double? value,
-  ) =>
+  bool? getBool(String? key) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #setDouble,
-          [
-            key,
-            value,
-          ],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
+            Invocation.method(#getBool, [key]),
+            returnValueForMissingStub: null,
+          )
+          as bool?);
   @override
-  _i4.Future<bool> setString(
-    String? key,
-    String? value,
-  ) =>
+  int? getInt(String? key) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #setString,
-          [
-            key,
-            value,
-          ],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
+            Invocation.method(#getInt, [key]),
+            returnValueForMissingStub: null,
+          )
+          as int?);
   @override
-  _i4.Future<bool> setStringList(
-    String? key,
-    List<String>? value,
-  ) =>
+  double? getDouble(String? key) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #setStringList,
-          [
-            key,
-            value,
-          ],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
+            Invocation.method(#getDouble, [key]),
+            returnValueForMissingStub: null,
+          )
+          as double?);
   @override
-  _i4.Future<bool> remove(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #remove,
-          [key],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
+  String? getString(String? key) =>
+      (super.noSuchMethod(
+            Invocation.method(#getString, [key]),
+            returnValueForMissingStub: null,
+          )
+          as String?);
   @override
-  _i4.Future<bool> commit() => (super.noSuchMethod(
-        Invocation.method(
-          #commit,
-          [],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
+  bool containsKey(String? key) =>
+      (super.noSuchMethod(
+            Invocation.method(#containsKey, [key]),
+            returnValue: false,
+            returnValueForMissingStub: false,
+          )
+          as bool);
   @override
-  _i4.Future<bool> clear() => (super.noSuchMethod(
-        Invocation.method(
-          #clear,
-          [],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
+  List<String>? getStringList(String? key) =>
+      (super.noSuchMethod(
+            Invocation.method(#getStringList, [key]),
+            returnValueForMissingStub: null,
+          )
+          as List<String>?);
   @override
-  _i4.Future<void> reload() => (super.noSuchMethod(
-        Invocation.method(
-          #reload,
-          [],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
+  _i4.Future<bool> setBool(String? key, bool? value) =>
+      (super.noSuchMethod(
+            Invocation.method(#setBool, [key, value]),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+  @override
+  _i4.Future<bool> setInt(String? key, int? value) =>
+      (super.noSuchMethod(
+            Invocation.method(#setInt, [key, value]),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+  @override
+  _i4.Future<bool> setDouble(String? key, double? value) =>
+      (super.noSuchMethod(
+            Invocation.method(#setDouble, [key, value]),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+  @override
+  _i4.Future<bool> setString(String? key, String? value) =>
+      (super.noSuchMethod(
+            Invocation.method(#setString, [key, value]),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+  @override
+  _i4.Future<bool> setStringList(String? key, List<String>? value) =>
+      (super.noSuchMethod(
+            Invocation.method(#setStringList, [key, value]),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+  @override
+  _i4.Future<bool> remove(String? key) =>
+      (super.noSuchMethod(
+            Invocation.method(#remove, [key]),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+  @override
+  _i4.Future<bool> commit() =>
+      (super.noSuchMethod(
+            Invocation.method(#commit, []),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+  @override
+  _i4.Future<bool> clear() =>
+      (super.noSuchMethod(
+            Invocation.method(#clear, []),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+  @override
+  _i4.Future<void> reload() =>
+      (super.noSuchMethod(
+            Invocation.method(#reload, []),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
 }
 
 /// A class which mocks [PageController].
@@ -231,71 +175,91 @@ class MockSharedPreferences extends _i1.Mock implements _i3.SharedPreferences {
 /// See the documentation for Mockito's code generation for more information.
 class MockPageController extends _i1.Mock implements _i5.PageController {
   @override
-  int get initialPage => (super.noSuchMethod(
-        Invocation.getter(#initialPage),
-        returnValue: 0,
-        returnValueForMissingStub: 0,
-      ) as int);
+  int get initialPage =>
+      (super.noSuchMethod(
+            Invocation.getter(#initialPage),
+            returnValue: 0,
+            returnValueForMissingStub: 0,
+          )
+          as int);
   @override
-  bool get keepPage => (super.noSuchMethod(
-        Invocation.getter(#keepPage),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
+  bool get keepPage =>
+      (super.noSuchMethod(
+            Invocation.getter(#keepPage),
+            returnValue: false,
+            returnValueForMissingStub: false,
+          )
+          as bool);
   @override
-  double get viewportFraction => (super.noSuchMethod(
-        Invocation.getter(#viewportFraction),
-        returnValue: 0.0,
-        returnValueForMissingStub: 0.0,
-      ) as double);
+  double get viewportFraction =>
+      (super.noSuchMethod(
+            Invocation.getter(#viewportFraction),
+            returnValue: 0.0,
+            returnValueForMissingStub: 0.0,
+          )
+          as double);
   @override
-  bool get keepScrollOffset => (super.noSuchMethod(
-        Invocation.getter(#keepScrollOffset),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
+  bool get keepScrollOffset =>
+      (super.noSuchMethod(
+            Invocation.getter(#keepScrollOffset),
+            returnValue: false,
+            returnValueForMissingStub: false,
+          )
+          as bool);
   @override
-  double get initialScrollOffset => (super.noSuchMethod(
-        Invocation.getter(#initialScrollOffset),
-        returnValue: 0.0,
-        returnValueForMissingStub: 0.0,
-      ) as double);
+  double get initialScrollOffset =>
+      (super.noSuchMethod(
+            Invocation.getter(#initialScrollOffset),
+            returnValue: 0.0,
+            returnValueForMissingStub: 0.0,
+          )
+          as double);
   @override
-  Iterable<_i2.ScrollPosition> get positions => (super.noSuchMethod(
-        Invocation.getter(#positions),
-        returnValue: <_i2.ScrollPosition>[],
-        returnValueForMissingStub: <_i2.ScrollPosition>[],
-      ) as Iterable<_i2.ScrollPosition>);
+  Iterable<_i2.ScrollPosition> get positions =>
+      (super.noSuchMethod(
+            Invocation.getter(#positions),
+            returnValue: <_i2.ScrollPosition>[],
+            returnValueForMissingStub: <_i2.ScrollPosition>[],
+          )
+          as Iterable<_i2.ScrollPosition>);
   @override
-  bool get hasClients => (super.noSuchMethod(
-        Invocation.getter(#hasClients),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
+  bool get hasClients =>
+      (super.noSuchMethod(
+            Invocation.getter(#hasClients),
+            returnValue: false,
+            returnValueForMissingStub: false,
+          )
+          as bool);
   @override
-  _i2.ScrollPosition get position => (super.noSuchMethod(
-        Invocation.getter(#position),
-        returnValue: _FakeScrollPosition_0(
-          this,
-          Invocation.getter(#position),
-        ),
-        returnValueForMissingStub: _FakeScrollPosition_0(
-          this,
-          Invocation.getter(#position),
-        ),
-      ) as _i2.ScrollPosition);
+  _i2.ScrollPosition get position =>
+      (super.noSuchMethod(
+            Invocation.getter(#position),
+            returnValue: _FakeScrollPosition_0(
+              this,
+              Invocation.getter(#position),
+            ),
+            returnValueForMissingStub: _FakeScrollPosition_0(
+              this,
+              Invocation.getter(#position),
+            ),
+          )
+          as _i2.ScrollPosition);
   @override
-  double get offset => (super.noSuchMethod(
-        Invocation.getter(#offset),
-        returnValue: 0.0,
-        returnValueForMissingStub: 0.0,
-      ) as double);
+  double get offset =>
+      (super.noSuchMethod(
+            Invocation.getter(#offset),
+            returnValue: 0.0,
+            returnValueForMissingStub: 0.0,
+          )
+          as double);
   @override
-  bool get hasListeners => (super.noSuchMethod(
-        Invocation.getter(#hasListeners),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
+  bool get hasListeners =>
+      (super.noSuchMethod(
+            Invocation.getter(#hasListeners),
+            returnValue: false,
+            returnValueForMissingStub: false,
+          )
+          as bool);
   @override
   _i4.Future<void> animateToPage(
     int? page, {
@@ -303,59 +267,48 @@ class MockPageController extends _i1.Mock implements _i5.PageController {
     required _i6.Curve? curve,
   }) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #animateToPage,
-          [page],
-          {
-            #duration: duration,
-            #curve: curve,
-          },
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
+            Invocation.method(
+              #animateToPage,
+              [page],
+              {#duration: duration, #curve: curve},
+            ),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
   @override
   void jumpToPage(int? page) => super.noSuchMethod(
-        Invocation.method(
-          #jumpToPage,
-          [page],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#jumpToPage, [page]),
+    returnValueForMissingStub: null,
+  );
   @override
   _i4.Future<void> nextPage({
     required Duration? duration,
     required _i6.Curve? curve,
   }) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #nextPage,
-          [],
-          {
-            #duration: duration,
-            #curve: curve,
-          },
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
+            Invocation.method(#nextPage, [], {
+              #duration: duration,
+              #curve: curve,
+            }),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
   @override
   _i4.Future<void> previousPage({
     required Duration? duration,
     required _i6.Curve? curve,
   }) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #previousPage,
-          [],
-          {
-            #duration: duration,
-            #curve: curve,
-          },
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
+            Invocation.method(#previousPage, [], {
+              #duration: duration,
+              #curve: curve,
+            }),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
   @override
   _i2.ScrollPosition createScrollPosition(
     _i7.ScrollPhysics? physics,
@@ -363,45 +316,34 @@ class MockPageController extends _i1.Mock implements _i5.PageController {
     _i2.ScrollPosition? oldPosition,
   ) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #createScrollPosition,
-          [
-            physics,
-            context,
-            oldPosition,
-          ],
-        ),
-        returnValue: _FakeScrollPosition_0(
-          this,
-          Invocation.method(
-            #createScrollPosition,
-            [
+            Invocation.method(#createScrollPosition, [
               physics,
               context,
               oldPosition,
-            ],
-          ),
-        ),
-        returnValueForMissingStub: _FakeScrollPosition_0(
-          this,
-          Invocation.method(
-            #createScrollPosition,
-            [
-              physics,
-              context,
-              oldPosition,
-            ],
-          ),
-        ),
-      ) as _i2.ScrollPosition);
+            ]),
+            returnValue: _FakeScrollPosition_0(
+              this,
+              Invocation.method(#createScrollPosition, [
+                physics,
+                context,
+                oldPosition,
+              ]),
+            ),
+            returnValueForMissingStub: _FakeScrollPosition_0(
+              this,
+              Invocation.method(#createScrollPosition, [
+                physics,
+                context,
+                oldPosition,
+              ]),
+            ),
+          )
+          as _i2.ScrollPosition);
   @override
   void attach(_i2.ScrollPosition? position) => super.noSuchMethod(
-        Invocation.method(
-          #attach,
-          [position],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#attach, [position]),
+    returnValueForMissingStub: null,
+  );
   @override
   _i4.Future<void> animateTo(
     double? offset, {
@@ -409,71 +351,48 @@ class MockPageController extends _i1.Mock implements _i5.PageController {
     required _i6.Curve? curve,
   }) =>
       (super.noSuchMethod(
-        Invocation.method(
-          #animateTo,
-          [offset],
-          {
-            #duration: duration,
-            #curve: curve,
-          },
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
+            Invocation.method(
+              #animateTo,
+              [offset],
+              {#duration: duration, #curve: curve},
+            ),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
   @override
   void jumpTo(double? value) => super.noSuchMethod(
-        Invocation.method(
-          #jumpTo,
-          [value],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#jumpTo, [value]),
+    returnValueForMissingStub: null,
+  );
   @override
   void detach(_i2.ScrollPosition? position) => super.noSuchMethod(
-        Invocation.method(
-          #detach,
-          [position],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#detach, [position]),
+    returnValueForMissingStub: null,
+  );
   @override
   void dispose() => super.noSuchMethod(
-        Invocation.method(
-          #dispose,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#dispose, []),
+    returnValueForMissingStub: null,
+  );
   @override
   void debugFillDescription(List<String>? description) => super.noSuchMethod(
-        Invocation.method(
-          #debugFillDescription,
-          [description],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#debugFillDescription, [description]),
+    returnValueForMissingStub: null,
+  );
   @override
   void addListener(_i9.VoidCallback? listener) => super.noSuchMethod(
-        Invocation.method(
-          #addListener,
-          [listener],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#addListener, [listener]),
+    returnValueForMissingStub: null,
+  );
   @override
   void removeListener(_i9.VoidCallback? listener) => super.noSuchMethod(
-        Invocation.method(
-          #removeListener,
-          [listener],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#removeListener, [listener]),
+    returnValueForMissingStub: null,
+  );
   @override
   void notifyListeners() => super.noSuchMethod(
-        Invocation.method(
-          #notifyListeners,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
+    Invocation.method(#notifyListeners, []),
+    returnValueForMissingStub: null,
+  );
 }
